rxkad: Declare 'outsum' as a u_long
Building src/rxkad/test/stress_c.c currently fails to build on 32-bit
x86:

    stress_c.c: In function 'Copious':
    stress_c.c:171:39: error: passing argument 2 of 'EndRXKST_Copious' from incompatible pointer type [-Wincompatible-pointer-types]
      171 |         code = EndRXKST_Copious(call, &outsum);
          |                                       ^~~~~~~
          |                                       |
          |                                       size_t * {aka unsigned int *}
    In file included from stress_c.c:26:
    stress.h:100:26: note: expected 'u_long *' {aka 'long unsigned int *'} but argument is of type 'size_t *' {aka 'unsigned int *'}
      100 |         /*OUT*/ u_long * outsum);
          |                 ~~~~~~~~~^~~~~~
    make[3]: *** [.../src/config/Makefile.lwp:9: stress_c.o] Error 1

Change 'outsum' to be a u_long instead of a size_t, since the output
argument of RXKST_Copious() is defined as a u_long in the xg definition.

Commit d17adc3002 (rxkad: Make the test suite build again) originally
fixed this by changing 'outsum' from a 'long' to a 'size_t', which
happens to be the same size and signedness on most modern platforms. But
they are different sizes these days on some platforms, such as Debian
i386.

Reviewed-on: https://gerrit.openafs.org/16920
Tested-by: BuildBot <buildbot@rampaginggeek.com>

+++ src/rxkad/test/stress_c.c
@@ -117,7 +117,7 @@ Copious(struct client *c, char *buf, u_long buflen)
     long outlen = c->recvLen;
     long d = 23;
     long mysum;
-    size_t outsum;
+    u_long outsum;
 
     mysum = 0;
     for (i = 0; i < inlen; i++)
